pyproject.toml
[tool.poetry]
name = "bigquery-data-profiler"
version = "0.0.1"
description = "Data profiler for BigQuery"
authors = ["Attsun1031"]
license = "MIT"

[tool.poetry.dependencies]
python = "3.9.9"
google-cloud-bigquery = "^2.30.1"
typer = "^0.4.0"
pydantic = "^1.8.2"
PyYAML = "^6.0"
loguru = "^0.5.3"

[tool.poetry.dev-dependencies]
yapf = "^0.31.0"
pylint = "^2.11.1"
isort = "^5.10.1"
mypy = "^0.910"
types-PyYAML = "^6.0.1"
pytest = "^6.2.5"

[tool.yapf]
based_on_style = "google"
column_limit = 120

[tool.isort]
line_length = 120

[tool.pylint.MASTER]
ignore = "tests"

[tool.pylint.basic]
# Regular expression which should only match correct argument names
argument-rgx = "[a-z0-9_]{1,30}$"
# Regular expression which should only match correct variable names
variable-rgx = "[a-z0-9_]{1,30}$"
max-args = 8
min-similarity-lines = 10

[tool.pylint.format]
max-line-length = 120

[tool.pylint."MESSAGES CONTROL"]
disable = "R0201, R0402, R0903, W0511, W0703, C0114"

[tool.mypy]
ignore_missing_imports = true
follow_imports = "skip"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
